// src/app/models/weather.model.ts
export enum Weather {
    ClearSky = 'Clear Sky',
    Overcast = 'Overcast',
    Rain = 'Rain',
    HeavyRain = 'Heavy Rain',
    Thunderstorm = 'Thunderstorm',
    Fog = 'Fog',
    Snowfall = 'Snowfall',
    Blizzard = 'Blizzard',
    Windy = 'Windy',
    Gale = 'Gale',
    Hail = 'Hail',
    Heatwave = 'Heatwave',
    ColdSnap = 'Cold Snap',
    SandStorm = 'Sand Storm',
    TropicalStorm = 'Tropical Storm',
    EtherealStorm = 'Ethereal Storm',
    AcidRain = 'Acid Rain',
    LightningStorm = 'Lightning Storm',
    PurpleRain = 'Purple Rain',
    MeteorShower = 'Meteor Shower'
}
  